// packages/prisma-oppulence/generated/zod/inputTypeSchemas/PipelineStageUpdateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { NullableStringFieldUpdateOperationsInputSchema } from './NullableStringFieldUpdateOperationsInputSchema';
import { IntFieldUpdateOperationsInputSchema } from './IntFieldUpdateOperationsInputSchema';
import { NullableFloatFieldUpdateOperationsInputSchema } from './NullableFloatFieldUpdateOperationsInputSchema';
import { BoolFieldUpdateOperationsInputSchema } from './BoolFieldUpdateOperationsInputSchema';
import { NullableIntFieldUpdateOperationsInputSchema } from './NullableIntFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { PipelineUpdateOneRequiredWithoutStagesNestedInputSchema } from './PipelineUpdateOneRequiredWithoutStagesNestedInputSchema';
import { DealUpdateManyWithoutStageNestedInputSchema } from './DealUpdateManyWithoutStageNestedInputSchema';
import { StageRequiredFieldUpdateManyWithoutStageNestedInputSchema } from './StageRequiredFieldUpdateManyWithoutStageNestedInputSchema';
import { StageDurationUpdateOneWithoutStageNestedInputSchema } from './StageDurationUpdateOneWithoutStageNestedInputSchema';

export const PipelineStageUpdateInputSchema: z.ZodType<Prisma.PipelineStageUpdateInput> = z.object({
  id: z.union([ z.string().cuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  name: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  description: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  order: z.union([ z.number().int(),z.lazy(() => IntFieldUpdateOperationsInputSchema) ]).optional(),
  probability: z.union([ z.number(),z.lazy(() => NullableFloatFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  color: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  icon: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  isDefault: z.union([ z.boolean(),z.lazy(() => BoolFieldUpdateOperationsInputSchema) ]).optional(),
  isFinal: z.union([ z.boolean(),z.lazy(() => BoolFieldUpdateOperationsInputSchema) ]).optional(),
  isWon: z.union([ z.boolean(),z.lazy(() => BoolFieldUpdateOperationsInputSchema) ]).optional(),
  isLost: z.union([ z.boolean(),z.lazy(() => BoolFieldUpdateOperationsInputSchema) ]).optional(),
  autoMoveAfterDays: z.union([ z.number().int(),z.lazy(() => NullableIntFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  nextStageId: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  pipeline: z.lazy(() => PipelineUpdateOneRequiredWithoutStagesNestedInputSchema).optional(),
  deals: z.lazy(() => DealUpdateManyWithoutStageNestedInputSchema).optional(),
  requiredFields: z.lazy(() => StageRequiredFieldUpdateManyWithoutStageNestedInputSchema).optional(),
  stageDuration: z.lazy(() => StageDurationUpdateOneWithoutStageNestedInputSchema).optional()
}).strict();

export default PipelineStageUpdateInputSchema;